+++ ide/src/rust/ide/view/graph-editor/src/builtin/visualization/java_script/histogram.js
@@ -24,17 +24,17 @@ const BUTTON_HEIGHT = 25
  * A d3.js histogram visualization.
  *
  *
- * Data format (json):
+ * Data format (JSON):
  {
    "axis" : {
       "x" : { "label" : "x-axis label", "scale" : "linear" },
       "y" : { "label" : "y-axis label", "scale" : "logarithmic" },
    },
-   "focus" { "x" : 1.7, "y" : 2.1, "zoom" : 3.0 },
-   "color" :  "rgb(1.0,0.0,0.0)" },
-   "bins" : 10,
-   "data" : [
-     "values" : [0.1, 0.2, 0.1, 0.15, 0.7],
+   "focus" : { "x" : 1.7, "y" : 2.1, "zoom" : 3.0 },
+   "color" : "rgb(1.0,0.0,0.0)",
+   "bins"  : 10,
+   "data"  : [
+      "values" : [0.1, 0.2, 0.1, 0.15, 0.7],
    ]
  }
  */
@@ -230,35 +230,45 @@ class Histogram extends Visualization {
      */
     initPanAndZoom() {
         const canvas = this.canvas
-        const zoomClass = 'zoom'
         const minScale = 0.5
         const maxScale = 20
+        const rightButton = 2
+        const midButton = 1
+        const scrollWheel = 0
         const extent = [minScale, maxScale]
+        let startPos
         const zoom = d3
             .zoom()
             .filter(function () {
-                let right_button = 2
-                let mid_button = 1
-                let scroll_wheel = 0
                 switch (d3.event.type) {
                     case 'mousedown':
-                        return d3.event.button === right_button || d3.event.button === mid_button
+                        return d3.event.button === rightButton || d3.event.button === midButton
                     case 'wheel':
-                        return d3.event.button === scroll_wheel
+                        return d3.event.button === scrollWheel
                     default:
                         return false
                 }
+            })
+            .wheelDelta(function () {
+                const event = d3.event
+                const minDelta = 0.002
+                const medDelta = 0.05
+                const maxDelta = 1
+                const wheelSpeedMultiplier =
+                    event.deltaMode === 1 ? medDelta : event.deltaMode ? maxDelta : minDelta
+                return -event.deltaY * wheelSpeedMultiplier
             })
             .scaleExtent(extent)
             .extent([
                 [0, 0],
                 [canvas.inner.width, canvas.inner.height],
             ])
-            .on(zoomClass, zoomed)
+            .on('zoom', zoomed)
+            .on('start', startZoom)
 
         const zoomElem = this.svg
             .append('g')
-            .attr('class', zoomClass)
+            .attr('class', 'zoom')
             .attr('width', canvas.inner.width)
             .attr('height', canvas.inner.height)
             .style('fill', 'none')
@@ -266,17 +276,82 @@ class Histogram extends Visualization {
             .call(zoom)
 
         const self = this
+
+        let transformedScale = Object.assign({}, self.scale)
+
         /**
          * Helper function called on pan/scroll.
          */
         function zoomed() {
-            self.scale.zoom = d3.event.transform.k
-            let tmpScale = Object.assign({}, self.scale)
-            tmpScale.x = d3.event.transform.rescaleX(self.scale.x)
-            self.rescale(tmpScale, false)
+            if (d3.event.sourceEvent != null && d3.event.sourceEvent.buttons === rightButton) {
+                const rmbDivider = 5000.0
+                const zoomAmount = rmbZoomValue(d3.event.sourceEvent) / rmbDivider
+                const scale = Math.exp(zoomAmount)
+                const focus = startPos
+                const distanceScale = d3.zoomIdentity
+                    .translate(focus.x - (Y_AXIS_LABEL_WIDTH + MARGIN), focus.y - MARGIN)
+                    .scale(scale)
+                    .translate(-focus.x + (Y_AXIS_LABEL_WIDTH + MARGIN), -focus.y + MARGIN)
+                transformedScale.x = distanceScale.rescaleX(transformedScale.x)
+                transformedScale.zoom = transformedScale.zoom * scale
+            } else if (d3.event.sourceEvent != null && d3.event.sourceEvent.type === 'wheel') {
+                if (d3.event.sourceEvent.ctrlKey) {
+                    const pinchDivider = 100.0
+                    const zoomAmount = -d3.event.sourceEvent.deltaY / pinchDivider
+                    const scale = Math.exp(zoomAmount)
+                    const focus = startPos
+                    const distanceScale = d3.zoomIdentity
+                        .translate(focus.x - (Y_AXIS_LABEL_WIDTH + MARGIN), focus.y - MARGIN)
+                        .scale(scale)
+                        .translate(-focus.x + (Y_AXIS_LABEL_WIDTH + MARGIN), -focus.y + MARGIN)
+                    transformedScale.x = distanceScale.rescaleX(transformedScale.x)
+                    transformedScale.zoom = transformedScale.zoom * scale
+                } else {
+                    const distanceScale = d3.zoomIdentity.translate(
+                        -d3.event.sourceEvent.deltaX,
+                        -d3.event.sourceEvent.deltaY
+                    )
+                    transformedScale.x = distanceScale.rescaleX(transformedScale.x)
+                    if (distanceScale.rescaleY(transformedScale.y).domain()[0] >= 0) {
+                        transformedScale.y = distanceScale.rescaleY(transformedScale.y)
+                    }
+                }
+            } else {
+                transformedScale.x = d3.event.transform.rescaleX(transformedScale.x)
+                if (d3.event.transform.rescaleY(transformedScale.y).domain()[0] >= 0) {
+                    transformedScale.y = d3.event.transform.rescaleY(transformedScale.y)
+                }
+            }
+
+            self.rescale(transformedScale, false)
         }
 
-        return { zoomElem, zoom }
+        /**
+         * Return the position of this event in local canvas coordinates.
+         */
+        function getPos(event) {
+            return { x: event.offsetX, y: event.offsetY }
+        }
+
+        /**
+         * Return the zoom value computed from the initial right-mouse-button event to the current
+         * right-mouse event.
+         */
+        function rmbZoomValue(event) {
+            const end = getPos(event)
+            const dX = end.x - startPos.x
+            const dY = end.y - startPos.y
+            return dX - dY
+        }
+
+        /**
+         * Helper function called when starting to pan/scroll.
+         */
+        function startZoom() {
+            startPos = getPos(d3.event.sourceEvent)
+        }
+
+        return { zoomElem, zoom, transformedScale }
     }
 
     /**
@@ -310,14 +385,14 @@ class Histogram extends Visualization {
          * Section "Brushing for zooming".
          */
         const zoomIn = () => {
-            const xMin = self.scale.x.invert(extent[0])
-            const xMax = self.scale.x.invert(extent[1])
+            const xMin = zoom.transformedScale.x.invert(extent[0])
+            const xMax = zoom.transformedScale.x.invert(extent[1])
 
-            self.scale.x.domain([xMin, xMax])
+            zoom.transformedScale.x.domain([xMin, xMax])
             const dx = extent[1] - extent[0]
-            self.scale.zoom = self.scale.zoom * (self.canvas.inner.width / dx)
+            zoom.transformedScale.zoom = zoom.transformedScale.zoom * (self.canvas.inner.width / dx)
 
-            self.rescale(self.scale, true)
+            self.rescale(zoom.transformedScale, true)
         }
 
         const zoomInKeyEvent = event => {
@@ -355,16 +430,17 @@ class Histogram extends Visualization {
     /**
      * Helper function for rescaling the data points with a new scale.
      */
-    rescale(scale, with_animation) {
-        const animation_duration = with_animation ? ANIMATION_DURATION : 0.0
+    rescale(scale, withAnimation) {
+        const duration = withAnimation ? ANIMATION_DURATION : 0.0
         this.xAxis
             .transition()
-            .duration(animation_duration)
+            .duration(duration)
             .call(d3.axisBottom(scale.x).ticks(this.binCount()))
+        this.yAxis.transition().duration(duration).call(d3.axisLeft(scale.y))
         this.plot
             .selectAll('rect')
             .transition()
-            .duration(animation_duration)
+            .duration(duration)
             .attr(
                 'transform',
                 d =>
@@ -389,19 +465,19 @@ class Histogram extends Visualization {
         const dataPoints = this.data()
         const focus = this.focus()
 
-        let domain_x = [
+        let domainX = [
             extremesAndDeltas.xMin - extremesAndDeltas.paddingX,
             extremesAndDeltas.xMax + extremesAndDeltas.paddingX,
         ]
 
         if (focus !== undefined) {
             if (focus.x !== undefined && focus.zoom !== undefined) {
-                let padding_x = extremesAndDeltas.dx * (1 / (2 * focus.zoom))
-                domain_x = [focus.x - padding_x, focus.x + padding_x]
+                let paddingX = extremesAndDeltas.dx * (1 / (2 * focus.zoom))
+                domainX = [focus.x - paddingX, focus.x + paddingX]
             }
         }
 
-        const x = d3.scaleLinear().domain(domain_x).range([0, this.canvas.inner.width])
+        const x = d3.scaleLinear().domain(domainX).range([0, this.canvas.inner.width])
 
         this.xAxis
             .attr('transform', 'translate(0,' + this.canvas.inner.height + ')')
@@ -649,26 +725,27 @@ class Histogram extends Visualization {
         btn.appendChild(text)
 
         const self = this
-        const reset_zoom_and_pan = () => {
+        const resetZoomAndPan = () => {
             zoom.zoomElem.transition().duration(0).call(zoom.zoom.transform, d3.zoomIdentity)
 
-            let domain_x = [
+            let domainX = [
                 extremesAndDeltas.xMin - extremesAndDeltas.paddingX,
                 extremesAndDeltas.xMax + extremesAndDeltas.paddingX,
             ]
 
-            self.scale.x.domain(domain_x)
+            zoom.transformedScale.x.domain(domainX)
             self.scale.zoom = 1.0
-            self.rescale(self.scale, true)
+            zoom.transformedScale.zoom = self.scale.zoom
+            self.rescale(zoom.transformedScale, true)
         }
 
         document.addEventListener('keydown', e => {
             if (shortcuts.showAll(e)) {
-                reset_zoom_and_pan()
+                resetZoomAndPan()
             }
         })
 
-        btn.addEventListener('click', reset_zoom_and_pan)
+        btn.addEventListener('click', resetZoomAndPan)
         this.dom.appendChild(btn)
     }